feat: move localstorage load to create hook
