+++ packages/api-client/src/views/Request/RequestSection/RequestSection.vue
@@ -18,6 +18,7 @@ const {
   requestMutators,
 } = useWorkspace()
 
+console.log(activeRequest, 'ASDASDSD')
 const bodyMethods = ['POST', 'PUT', 'PATCH', 'DELETE']
 
 const sections = computed(() => {
